convert/server/pybackend/requirements.txt
flask
flask-cors
pdfplumber
python-docx
openai>=1.0.0
python-dotenv
